backend/requirements.txt
accelerate
boto3
botocore
certifi
fastapi
fastapi-health
fireworks-ai
google-api-core
google-auth
google_auth_oauthlib
google-cloud-core
json-repair
passlib
langchain
langchain-aws
langchain-anthropic
langchain-fireworks
langchain-community
langchain-core
langchain-experimental
langchain-google-vertexai
langchain-groq
langchain-openai
langchain-text-splitters
langchain-huggingface
langdetect
langsmith
langserve
neo4j-rust-ext
nltk
openai
psutil
pydantic
python-dotenv
python-magic
PyPDF2
PyMuPDF
starlette
sse-starlette
starlette-session
tqdm
unstructured
unstructured-client
unstructured-inference
urllib3
uvicorn
gunicorn
wikipedia
wrapt
yarl
youtube-transcript-api
zipp
sentence-transformers
google-cloud-logging
pypandoc
graphdatascience
Secweb
ragas
rouge_score
langchain-neo4j
pypandoc-binary
chardet
cryptography
sqlalchemy
psycopg2-binary
PyJWT
alembic
bcrypt

